third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryServices/blocklisted-services-with-uuid.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return getHIDDevice({
      filters: [{services: ['device_information']}],
      optionalServices: ['human_interface_device']})
    .then(({device}) => assert_promise_rejects_with_message(
      device.gatt.getPrimaryServices('human_interface_device'),
      new DOMException('Origin is not allowed to access the service. ' +
                       'Tip: Add the service UUID to \'optionalServices\' ' +
                       'in requestDevice() options. https://goo.gl/HxfxSQ',
      'SecurityError')));
}, 'Request for services. Does not return blocklisted service.');
</script>
